Konfigurierbare Ligen ...

+++ app.js
@@ -1,31 +1,31 @@
  'use strict';
 
-var uid = require(__dirname + "/api/helpers/UID");
-var crypt = require(__dirname + "/api/helpers/Crypt");
-var start = (new Date()).getTime();
-var mkdirp = require('mkdirp');
-var path = require('path');
-var fs = require('fs-extra');
-var config = require('config');
-var util = require('util');
-var moment = require('moment');
-var logDirectory = "";
-var jsonfile = require('jsonfile');
+let uid = require(`${__dirname}/api/helpers/UID`);
+let crypt = require(`${__dirname}/api/helpers/Crypt`);
+let start = (new Date()).getTime();
+let mkdirp = require('mkdirp');
+let path = require('path');
+let fs = require('fs-extra');
+let config = require('config');
+let util = require('util');
+let moment = require('moment');
+let logDirectory = "";
+let jsonfile = require('jsonfile');
 jsonfile.spaces = 4;
 
 _init();
-var SwaggerExpress = require('swagger-express-mw');
-var express = require('express');
-var app = new express();
-var https = require('https');
-var http = require('http');
-var logger = require(__dirname + "/api/helpers/Logger");
-var morgan = require('morgan');
-var cors = require('cors');
-var cookieParser = require('cookie-parser');
+let SwaggerExpress = require('swagger-express-mw');
+let express = require('express');
+let app = new express();
+let https = require('https');
+let http = require('http');
+let logger = require(`${__dirname}/api/helpers/Logger`);
+let morgan = require('morgan');
+let cors = require('cors');
+let cookieParser = require('cookie-parser');
 module.exports = app; // for testing
 
-var serveIndex = require('serve-index');
+let serveIndex = require('serve-index');
 
 function logAll(message) {
     logger.console.info(message);
@@ -33,15 +33,15 @@ function logAll(message) {
 }
 
 function FNCheckConfig() {
-    var configFile = path.resolve(this.target + this.file);
-    var oConfig = (fs.existsSync(configFile)) ? jsonfile.readFileSync(configFile) : {};
+    let configFile = path.resolve(this.target + this.file);
+    let oConfig = (fs.existsSync(configFile)) ? jsonfile.readFileSync(configFile) : {};
     if (!oConfig.secret) {
         oConfig.secret = uid.generate(true);
         jsonfile.writeFileSync(configFile, oConfig);
     }
 
     if (!oConfig[config.get("bedelos.adminuser")]) {
-        var adminuser = config.get("bedelos.adminuser");
+        let adminuser = config.get("bedelos.adminuser");
         throw new Error("No password for " + adminuser + " found in data/config/config.json. Call 'node adminSecret.js' to set it.");
     }
 
@@ -49,14 +49,14 @@ function FNCheckConfig() {
 }
 
 function FNCheckUserStore() {
-    var userStoreFile = path.resolve(this.target + this.file);
+    let userStoreFile = path.resolve(this.target + this.file);
     if (!fs.existsSync(userStoreFile)) {
         jsonfile.writeFileSync(userStoreFile, {});
     }
 }
 
 function FNCheckAccessCounterStore() {
-    var accessCounterStoreFile = path.resolve(this.target + this.file);
+    let accessCounterStoreFile = path.resolve(this.target + this.file);
     if (!fs.existsSync(accessCounterStoreFile)) {
         jsonfile.writeFileSync(accessCounterStoreFile, {
             today: {
@@ -75,49 +75,53 @@ function FNCheckAccessCounterStore() {
 }
 
 function _init(){
-    var aItemsToResolve = [
-        { bFile: false, target: __dirname + "/" + config.get("log.dir") },
-        { bFile: false, target: __dirname + "/" + config.get("temp.dir") },
-        { bFile: false, target: __dirname + "/../backups" },
-        { bFile: false, target: __dirname + "/data" },
+    let aLigen = Object.keys(config.get('bedelos.ligen'));
 
-        { bFile: false, target: __dirname + "/data/config/" },
-        { bFile: true,  target: __dirname + "/data/config", file: "/config.json", after: FNCheckConfig },
-        { bFile: true,  target: __dirname + "/data/config", file: "/userStore.json", after: FNCheckUserStore },
-        { bFile: true,  target: __dirname + "/data/config", file: "/counter.json", after: FNCheckAccessCounterStore },
+    let aItemsToResolve = [
+        { bFile: false, target: `${__dirname}/${config.get("log.dir")}` },
+        { bFile: false, target: `${__dirname}/${config.get("temp.dir")}` },
+        { bFile: false, target: `${__dirname}/../backups}` },
+        { bFile: false, target: `${__dirname}/data}` },
 
-        { bFile: false, target: __dirname + "/data/saison/" },
-        { bFile: false, target: __dirname + "/data/saison/" + config.get("bedelos.saison") },
-        { bFile: false, target: __dirname + "/data/saison/" + config.get("bedelos.saison") + "/ergebnisse" },
-        { bFile: false, target: __dirname + "/data/saison/" + config.get("bedelos.saison") + "/statistiken" },
-        { bFile: false, target: __dirname + "/data/saison/" + config.get("bedelos.saison") + "/tabellen" },
-        { bFile: false, target: __dirname + "/data/saison/" + config.get("bedelos.saison") + "/pictures" },
-        { bFile: false, target: __dirname + "/data/saison/" + config.get("bedelos.saison") + "/inbox" },
+        { bFile: false, target: `${__dirname}/data/config/}` },
+        { bFile: true,  target: `${__dirname}/data/config`, file: "/config.json", after: FNCheckConfig },
+        { bFile: true,  target: `${__dirname}/data/config`, file: "/userStore.json", after: FNCheckUserStore },
+        { bFile: true,  target: `${__dirname}/data/config`, file: "/counter.json", after: FNCheckAccessCounterStore },
 
-        { bFile: true, target: __dirname + "/data/saison/" + config.get("bedelos.saison"), file: "/statistiken/klnord.json" },
-        { bFile: true, target: __dirname + "/data/saison/" + config.get("bedelos.saison"), file: "/statistiken/klsued.json" },
-        { bFile: true, target: __dirname + "/data/saison/" + config.get("bedelos.saison"), file: "/statistiken/bzLiga.json" },
-        { bFile: true, target: __dirname + "/data/saison/" + config.get("bedelos.saison"), file: "/statistiken/oberliga.json" },
-        { bFile: true, target: __dirname + "/data/saison/" + config.get("bedelos.saison"), file: "/tabellen/klnord.json" },
-        { bFile: true, target: __dirname + "/data/saison/" + config.get("bedelos.saison"), file: "/tabellen/klsued.json" },
-        { bFile: true, target: __dirname + "/data/saison/" + config.get("bedelos.saison"), file: "/tabellen/bzLiga.json" },
-        { bFile: true, target: __dirname + "/data/saison/" + config.get("bedelos.saison"), file: "/tabellen/oberliga.json" },
-        { bFile: true, target: __dirname + "/data/saison/" + config.get("bedelos.saison"), file: "/Spielplan.json" },
-        { bFile: true, target: __dirname + "/data/saison/" + config.get("bedelos.saison"), file: "/Teams.json" },
-        { bFile: true, target: __dirname + "/data/saison/" + config.get("bedelos.saison"), file: "/Wertung.json" }
+        { bFile: false, target: `${__dirname}/data/saison/}` },
+        { bFile: false, target: `${__dirname}/data/saison/${config.get("bedelos.saison")}` },
+        { bFile: false, target: `${__dirname}/data/saison/${config.get("bedelos.saison")}/ergebnisse` },
+        { bFile: false, target: `${__dirname}/data/saison/${config.get("bedelos.saison")}/statistiken` },
+        { bFile: false, target: `${__dirname}/data/saison/${config.get("bedelos.saison")}/tabellen` },
+        { bFile: false, target: `${__dirname}/data/saison/${config.get("bedelos.saison")}/pictures` },
+        { bFile: false, target: `${__dirname}/data/saison/${config.get("bedelos.saison")}/inbox` },
+
+        { bFile: true, target: `${__dirname}/data/saison/${config.get("bedelos.saison")}`, file: "/Spielplan.json" },
+        { bFile: true, target: `${__dirname}/data/saison/${config.get("bedelos.saison")}`, file: "/Teams.json" },
+        { bFile: true, target: `${__dirname}/data/saison/${config.get("bedelos.saison")}`, file: "/Vereine.json" },
+        { bFile: true, target: `${__dirname}/data/saison/${config.get("bedelos.saison")}`, file: "/Wertung.json" }
     ];
 
-    var defaultPath = path.resolve(__dirname + "/config/default/saison/xxyy");
+    aLigen.forEach(liga => {
+        aItemsToResolve.push({ bFile: true, target: `${__dirname}/data/saison/${config.get("bedelos.saison")}`, file: `/statistiken/${liga}.json` });
+        aItemsToResolve.push({ bFile: true, target: `${__dirname}/data/saison/${config.get("bedelos.saison")}`, file: `/tabellen/${liga}.json` });
+    });
 
-    for(var i = 0; i<aItemsToResolve.length; i++) {
+    let defaultPath = path.resolve(`${__dirname}/config/default/saison/xxyy`);
+
+    for(let i = 0; i<aItemsToResolve.length; i++) {
         if (aItemsToResolve[i].bFile) {
-            var sTargetFile = path.resolve(aItemsToResolve[i].target + aItemsToResolve[i].file);
+            let sTargetFile = path.resolve(aItemsToResolve[i].target + aItemsToResolve[i].file);
             if (!fs.existsSync(sTargetFile)) {
-                var sSourceFile = path.resolve(defaultPath + aItemsToResolve[i].file);
-                try { fs.copySync(sSourceFile, sTargetFile); } catch (err) {}
+                let sSourceFile = path.resolve(defaultPath + aItemsToResolve[i].file);
+                if (!fs.existsSync(sSourceFile)) {
+                    jsonfile.writeFileSync(sTargetFile, {});
+                } else {
+                    try { fs.copySync(sSourceFile, sTargetFile); } catch (err) {}
+                }
             }
         } else {
-            var sTargetDir = path.resolve(aItemsToResolve[i].target);
+            let sTargetDir = path.resolve(aItemsToResolve[i].target);
             if (!fs.existsSync(sTargetDir)) {
                 mkdirp.sync(sTargetDir);
             }
@@ -160,15 +164,15 @@ try {
         swaggerExpress.register(app);
 
         if (config.get("server.http")) {
-            var httpPort = config.get("server.http.port");
+            let httpPort = config.get("server.http.port");
             http.createServer(app).listen(httpPort);
             logAll(util.format("HTTP server running and listening on port=%d", httpPort));
         }
         // HTTPS service
         if (config.get("server.https")) {
-            var httpsPort = config.get("server.https.port");
-            var httpsKeyFile = config.get("server.https.keyFile");
-            var httpsCertFile = config.get("server.https.certFile");
+            let httpsPort = config.get("server.https.port");
+            let httpsKeyFile = config.get("server.https.keyFile");
+            let httpsCertFile = config.get("server.https.certFile");
             https.createServer({
                 key: fs.readFileSync(httpsKeyFile),
                 cert: fs.readFileSync(httpsCertFile)
@@ -176,7 +180,7 @@ try {
             logAll(util.format("HTTPS server running and listening on port=%d (key file=%s, certificate file=%s)",
                 httpsPort, httpsKeyFile, httpsCertFile));
         }
-        var took = (new Date()).getTime() - start;
+        let took = (new Date()).getTime() - start;
         logAll(util.format("Startup took %s milliseconds.", took));
 
     });
